backend/requirements.txt
flask
flask-cors
earthengine-api
google-api-python-client
google-cloud-documentai
gunicorn
